+++ docs/.vuepress/config.js
@@ -1,0 +1,19 @@
+module.exports = {
+    base: '/NativeCriteria/',
+    title: 'NativeCriteria',
+    description: 'Ultra lightweight library to generate dynamic SQL based on hibernate session',
+    ga: 'UA-19090015-3',
+    themeConfig: {
+        sidebar: [
+            ['/', "Introduction"],
+            ['/basic-examples.html', "Basic examples"],
+            ['/mappers.html', 'Mappers'],
+            ['/spring-data-integration.html', 'Spring Data integration'],
+            ['/logging.html', 'Logging'],
+            ['/release-notes.html', 'Release notes']
+        ]
+    },
+    markdown: {
+        lineNumbers: true
+    }
+};